backend/source/requirements.txt
numpy>=1.24.4
pillow>=9.4.0
flask>=3.0.3
flask_cors>=4.0.1
matplotlib>=3.9.1
opencv-python>=4.7.0